articular-examples: applied API improvements

+++ articular-examples/src/main/java/articular/example/labs/techdemos/jme/testjaime/components/InputComponent.java
@@ -47,6 +47,6 @@ public final class InputComponent implements Component {
     @Override
     public Id getId() {
         return new Component.Id((hashCode() >>> 16) ^
-                GameComponents.INPUT_COMPONENTS.getEntity().getId().intValue());
+                GameComponents.INPUT_COMPONENTS.getEntity().getId().longValue());
     }
 }
